- added ui error tests, README.md and final cleanup
