+++ .idea/deploymentTargetSelector.xml
@@ -5,9 +5,6 @@
       <SelectionState runConfigName="app">
         <option name="selectionMode" value="DROPDOWN" />
       </SelectionState>
-      <SelectionState runConfigName="givenProductListDisplayedWhenInputtingValidSearchQueryThenEmptySearchSectionIsShown()">
-        <option name="selectionMode" value="DROPDOWN" />
-      </SelectionState>
       <SelectionState runConfigName="givenProductListDisplayedWhenInputtingValidSearchQueryThenFilteredMultipleItemListIsShown()">
         <option name="selectionMode" value="DROPDOWN" />
       </SelectionState>
