+++ README.md
@@ -1,0 +1,16 @@
+# Instructions
+
+### 1. Setup
+
+In order to run this app you need to add to project root file `secure.properties` that have content below
+
+`baseUrl="https://mock.web.server.io/`
+
+
+### 1. App description
+
+- This app is build using H&B mock web server api to load list of products.
+- It shows list of Products and top search section, that filters the list or shows not found error
+- I used jetpack compose, hilt, Junit5, Mockk and Kotlin coroutines and flow.
+- App is split by the feature,
+- This code is WIP
