+++ app/src/androidTest/java/com/freemimp/hbtest/ProductListScreenTest.kt
@@ -5,16 +5,49 @@ import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performTextInput
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.freemimp.products.di.ViewModelModule
+import com.freemimp.products.domain.model.Product
+import com.freemimp.products.domain.repositories.ProductRepository
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
+import dagger.hilt.android.testing.UninstallModules
+import dagger.hilt.components.SingletonComponent
+import io.mockk.coEvery
+import io.mockk.mockk
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import javax.inject.Singleton
 
+@UninstallModules(ViewModelModule::class)
 @HiltAndroidTest
 @RunWith(AndroidJUnit4::class)
 class ProductListScreenTest {
+
+    @InstallIn(SingletonComponent::class)
+    @Module
+    object MockRepositoryModule {
+
+        @Singleton
+        @Provides
+        fun provideMockRepository(): ProductRepository {
+            return mockk<ProductRepository>(relaxed = true).apply {
+                coEvery { this@apply.getProducts() } returns Result.success((1..20).map {
+                    Product(
+                        sku = "$it",
+                        name = "Product $it",
+                        price = "$0.00",
+                        description = "Description$it",
+                        brand = "Brand$it",
+                    )
+                })
+            }
+        }
+    }
 
     @get:Rule(order = 0)
     var hiltRule = HiltAndroidRule(this)
