+++ app/src/androidTest/java/com/freemimp/hbtest/utils/TestException.kt
@@ -1,0 +1,3 @@
+package com.freemimp.hbtest.utils
+
+object TestException : Throwable("This is test exception")
