--- core/src/main/java/com/freemimp/core/model/Product.kt
@@ -1,9 +1,0 @@
-package com.freemimp.core.model
-
-data class Product(
-    val sku: String,
-    val name: String,
-    val price: Double,
-    val description: String,
-    val brand: String
-)
