--- core/src/main/java/com/freemimp/core/repositories/ProductRepository.kt
@@ -1,7 +1,0 @@
-package com.freemimp.core.repositories
-
-import com.freemimp.core.model.Product
-
-interface ProductRepository {
-    suspend fun getProducts(): List<Product>
-}
