+++ products/build.gradle.kts
@@ -55,9 +55,6 @@ android {
 }
 
 dependencies {
-    implementation(project(":core"))
-
-
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.lifecycle.runtime.ktx)
     implementation(libs.androidx.activity.compose)
